rename utils file and add test for isValidPlaylist method
