--- src/util.mjs
@@ -1,16 +1,0 @@
-import { ProtocolNotSupported } from './exceptions/ProtocolNotSupported.mjs';
-
-const isValidUrl = (url, protocols = ['http:', 'https:', 'ftp:', 'sftp:']) => {
-  try {
-    const { protocol = '' } = new URL(url);
-    if (protocol && !protocols.includes(`${protocol}`)) {
-      throw new ProtocolNotSupported(`${protocol} not supported. Supported protocols are ${protocols.join(', ')}`);
-    }
-
-    return true;
-  } catch (error) {
-    throw error;
-  }
-};
-
-export { isValidUrl };
